Fixed wrong extension metadata stored in cache when adding extension.
When adding (or updating) an extension the versions-tag was never cleared in the metadata stored in the cache. This led to different metadata at the backend and in the cache. Rewrote to never update the cache directly but instead I call the IndexExtension function when done. This updates the cache by reading the metadata actually found at the backend. The positive is that we always get the same result as when reindexing but the negative side is we need to fetch the metadata from the backend, when we already have it.